Safari support
For some reason safari uses the chrome password character ('•') but it's larger by default despite not being declared in the user agent stylesheet. To fix it I just found a number that works well to match the width of the larger character and declared a common difference to be sure the caret doesn't get misplaced.